Adjective co-occurrence

This data set includes counts for sentential co-occurrence of adjectives in the Brown corpus, based on Justeson & Katz (1990). The data is extracted from the same data set available in ?brown. The tags used to extract adjectives match "^jj.*".

Usage

adjective_cooccurrence

Format

A data frame with five variables, and two attributes: attribute corpus_size contains the number of tokens in the whole corpus attribute unique_jj contains the number of tokens tagged as adjectives (jj.*) in the whole corpus

word

first adjective

collocate

second adjective

o11

frequency of co-occurrence of word and collocate

f1

frequency of word

f2

frequency of collocate

References

Justeson, John S., and Slava M. Katz. "Co-occurrences of antonymous adjectives and their contexts." Computational linguistics 17.1 (1991): 1-20.

See also

Examples

data(adjective_cooccurrence)
str(adjective_cooccurrence)
#> 'data.frame':	56236 obs. of  5 variables:
#>  $ word     : chr  "grand" "executive" "possible" "hard-fought" ...
#>  $ collocate: chr  "recent" "over-all" "superior" "superior" ...
#>  $ o11      : int  1 1 2 1 1 3 1 3 1 1 ...
#>  $ f1       : int  45 15 374 2 2 45 45 978 32 22 ...
#>  $ f2       : int  179 35 45 45 374 978 29 29 4 4 ...
#>  - attr(*, "corpus_size")= int 1161192
#>  - attr(*, "unique_jj")= int 7955

attributes(adjective_cooccurrence)$unique_jj
#> [1] 7955